+++ pkg/process/util/orchestrator/cache_test.go
@@ -3,9 +3,9 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2020 Datadog, Inc.
 
-// +build kubeapiserver
+// +build orchestrator
 
-package apiserver
+package orchestrator
 
 import (
 	"testing"
